common: move boost_serialization_helper.h out and cleanup includes
