+++ src/cryptonote_core/tx_pool.cpp
@@ -40,7 +40,6 @@
 #include "blockchain.h"
 #include "blockchain_db/locked_txn.h"
 #include "blockchain_db/blockchain_db.h"
-#include "common/boost_serialization_helper.h"
 #include "int-util.h"
 #include "misc_language.h"
 #include "warnings.h"
