+++ tests/core_tests/chaingen_serialization.h
